refactor: share app and project service cores
