+++ src/deps/services/shared/fontLoader.js
@@ -1,0 +1,13 @@
+export const loadFont = async (fontName, fontUrl) => {
+  const existingFont = Array.from(document.fonts).find(
+    (font) => font.family === fontName,
+  );
+  if (existingFont) {
+    return existingFont;
+  }
+
+  const fontFace = new FontFace(fontName, `url(${fontUrl})`);
+  await fontFace.load();
+  document.fonts.add(fontFace);
+  return fontFace;
+};
