+++ src/deps/services/web/projectService.js
@@ -4,18 +4,10 @@ import {
   createInsiemeWebStoreAdapter,
   initializeProject as initializeWebProject,
 } from "../../infra/web/webRepositoryAdapter.js";
-import { createBundle } from "../../../utils/bundleUtils.js";
 import {
   createProjectCollabService,
   createWebSocketTransport,
 } from "../../../collab/v2/index.js";
-import {
-  getImageDimensions,
-  getVideoDimensions,
-  extractWaveformData,
-  extractVideoThumbnail,
-  detectFileType,
-} from "../../../utils/fileProcessors.js";
 import { projectRepositoryStateToDomainState } from "../../../domain/v2/stateProjection.js";
 import { createPersistedInMemoryClientStore } from "./collabClientStore.js";
 import {
@@ -28,25 +20,9 @@ import {
   applyCommandToRepository,
   assertRepositoryCommandEvent,
   assertV2State,
-  createProjectRepository,
   repositoryEventToCommand,
 } from "../shared/projectRepository.js";
-import { getOrCreateLocked } from "../shared/getOrCreateLocked.js";
-
-// Font loading helper
-const loadFont = async (fontName, fontUrl) => {
-  const existingFont = Array.from(document.fonts).find(
-    (font) => font.family === fontName,
-  );
-  if (existingFont) {
-    return existingFont;
-  }
-
-  const fontFace = new FontFace(fontName, `url(${fontUrl})`);
-  await fontFace.load();
-  document.fonts.add(fontFace);
-  return fontFace;
-};
+import { createBundle } from "../../../utils/bundleUtils.js";
 
 const countImageEntries = (imagesData) =>
   Object.values(imagesData?.items || {}).filter(
@@ -116,65 +92,16 @@ export const createProjectService = ({
     fn(`[routevn.collab.web] ${message}`, meta);
   };
 
-  // Repository cache
-  const repositoriesByProject = new Map();
-  const adaptersByProject = new Map();
   const collabApplyQueueByProject = new Map();
   const collabLastCommittedIdByProject = new Map();
   const collabAppliedCommittedIdsByProject = new Map();
 
-  // Initialization locks - prevents duplicate initialization
-  const initLocksByProject = new Map(); // projectId -> Promise<Repository>
-
-  // Current repository cache (for sync access after ensureRepository is called)
-  let currentRepository = null;
-  let currentProjectId = null;
-  let currentAdapter = null;
-
-  // Get current projectId from URL query params
-  const getCurrentProjectId = () => {
-    const { p } = router.getPayload();
-    return p;
-  };
-
-  const getProjectMetadataFromEntries = async (projectId) => {
-    if (!db || typeof db.get !== "function") {
-      return {
-        name: "",
-        description: "",
-      };
-    }
-
-    const entries = (await db.get("projectEntries")) || [];
-    const entry = Array.isArray(entries)
-      ? entries.find((item) => item?.id === projectId)
-      : null;
-    return {
-      name: entry?.name || "",
-      description: entry?.description || "",
-    };
-  };
-
-  const getAdapterByProject = async (projectId) => {
-    if (adaptersByProject.has(projectId)) {
-      return adaptersByProject.get(projectId);
-    }
-
-    const store = await createInsiemeWebStoreAdapter(projectId);
-    adaptersByProject.set(projectId, store);
-    return store;
-  };
-
-  const ensureCommittedIdLoaded = async (projectId) => {
+  const ensureCommittedIdLoaded = async (projectId, getStoreByProject) => {
     if (collabLastCommittedIdByProject.has(projectId)) {
       return collabLastCommittedIdByProject.get(projectId);
     }
-    const adapter = adaptersByProject.get(projectId) || null;
-    if (!adapter) {
-      collabLastCommittedIdByProject.set(projectId, 0);
-      return 0;
-    }
 
+    const adapter = await getStoreByProject(projectId);
     let loaded = 0;
     try {
       loaded = await loadCommittedCursor({ adapter, projectId });
@@ -189,16 +116,21 @@ export const createProjectService = ({
     return loaded;
   };
 
-  const persistCommittedCursor = async (projectId, cursor) => {
+  const persistCommittedCursor = async (
+    projectId,
+    cursor,
+    getStoreByProject,
+  ) => {
     const normalized = Number.isFinite(Number(cursor))
       ? Math.max(0, Math.floor(Number(cursor)))
       : 0;
-    const current = await ensureCommittedIdLoaded(projectId);
-    if (normalized <= current) return current;
+    const current = await ensureCommittedIdLoaded(projectId, getStoreByProject);
+    if (normalized <= current) {
+      return current;
+    }
 
     collabLastCommittedIdByProject.set(projectId, normalized);
-    const adapter = adaptersByProject.get(projectId) || null;
-    if (!adapter) return normalized;
+    const adapter = await getStoreByProject(projectId);
 
     try {
       await saveCommittedCursor({
@@ -216,6 +148,7 @@ export const createProjectService = ({
 
     return normalized;
   };
+
   const getAppliedCommittedSet = (projectId) => {
     let committedIds = collabAppliedCommittedIdsByProject.get(projectId);
     if (!committedIds) {
@@ -241,646 +174,89 @@ export const createProjectService = ({
     return next;
   };
 
-  const createSessionForProject = async ({
-    projectId,
-    token,
-    userId,
-    clientId,
-    endpointUrl,
-    partitions,
-    mode,
-    partitioning,
-  }) => {
-    collabLog("info", "create session requested", {
-      projectId,
-      endpointUrl: endpointUrl || null,
-      mode,
-      hasToken: Boolean(token),
-      userId,
-      clientId,
-    });
+  const storageAdapter = {
+    resolveProjectReferenceByProjectId: async ({ projectId }) => ({
+      cacheKey: projectId,
+      repositoryProjectId: projectId,
+    }),
 
-    const repository = await getRepositoryByProject(projectId);
-    const state = repository.getState();
-    assertV2State(state);
-    const adapter = adaptersByProject.get(projectId);
-    if (!adapter) {
-      throw new Error(`Store adapter not found for project '${projectId}'`);
-    }
-
-    const resolvedProjectId = state.project?.id || projectId;
-    const resolvedPartitions = partitioning.getBasePartitions(
-      resolvedProjectId,
-      partitions,
-    );
-    const projectMetadata = await getProjectMetadataFromEntries(projectId);
-    const clientStore = await createPersistedInMemoryClientStore({
-      projectId,
-      adapter,
-      logger: (entry) => {
-        if (entry?.level === "warn") {
-          collabLog("warn", "client store warning", entry);
-        }
-      },
-    });
-    await ensureCommittedIdLoaded(projectId);
-    const collabSession = createProjectCollabService({
-      projectId: resolvedProjectId,
-      projectName: projectMetadata.name,
-      projectDescription: projectMetadata.description,
-      initialState: projectRepositoryStateToDomainState({
-        repositoryState: state,
-        projectId: resolvedProjectId,
-      }),
-      token,
-      actor: {
-        userId,
-        clientId,
-      },
-      partitions: resolvedPartitions,
-      clientStore,
-      logger: (entry) => {
-        if (entry?.event === "connected") {
-          const globalLastCommittedId = Number(entry?.global_last_committed_id);
-          if (Number.isFinite(globalLastCommittedId)) {
-            collabLog("debug", "connected cursor", {
-              projectId: resolvedProjectId,
-              globalLastCommittedId,
-            });
-          }
-        }
-        collabLog("debug", "sync-client", entry);
-      },
-      onCommittedCommand: ({
-        command,
-        committedEvent,
-        sourceType,
-        isFromCurrentActor,
-      }) =>
-        queueCollabApply(projectId, async () => {
-          const applyCommittedEventToRepository = async ({
-            command,
-            committedEvent,
-            sourceType,
-            isFromCurrentActor,
-          }) => {
-            if (isFromCurrentActor) {
-              collabLog("debug", "command skipped (current actor source)", {
-                projectId,
-                sourceType,
-                commandId: command?.id || null,
-                commandType: command?.type || null,
-                committedId: Number.isFinite(
-                  Number(committedEvent?.committed_id),
-                )
-                  ? Number(committedEvent.committed_id)
-                  : null,
-              });
-              return;
-            }
-
-            const beforeState = repository.getState();
-            const beforeImagesCount = countImageEntries(beforeState?.images);
-            const applyResult = await applyCommandToRepository({
-              repository,
-              command,
-              projectId: resolvedProjectId,
-            });
-            if (applyResult.events.length === 0) {
-              collabLog("warn", "command ignored (no projection events)", {
-                projectId,
-                sourceType,
-                commandType: command?.type || null,
-                commandId: command?.id || null,
-              });
-              return;
-            }
-
-            const afterState = repository.getState();
-            const afterImagesCount = countImageEntries(afterState?.images);
-            collabLog("info", "remote command applied to repository", {
-              projectId,
-              commandType: command?.type || null,
-              applyMode: applyResult.mode,
-              eventCount: applyResult.events.length,
-              beforeImagesCount,
-              afterImagesCount,
-              sourceType,
-            });
-            if (typeof onRemoteEvent === "function") {
-              for (const event of applyResult.events) {
-                onRemoteEvent({
-                  projectId,
-                  sourceType,
-                  command,
-                  committedEvent,
-                  event,
-                });
-              }
-            }
-          };
-
-          const committedId = Number(committedEvent?.committed_id);
-          const hasCommittedId = Number.isFinite(committedId);
-          const lastCommittedId = await ensureCommittedIdLoaded(projectId);
-
-          if (hasCommittedId) {
-            const appliedCommittedIds = getAppliedCommittedSet(projectId);
-            if (appliedCommittedIds.has(committedId)) {
-              collabLog("debug", "committed event skipped (already applied)", {
-                projectId,
-                committedId,
-                sourceType,
-              });
-              return;
-            }
-            appliedCommittedIds.add(committedId);
-            if (appliedCommittedIds.size > 5000) {
-              const oldest = appliedCommittedIds.values().next().value;
-              appliedCommittedIds.delete(oldest);
-            }
-          }
-
-          await applyCommittedEventToRepository({
-            command,
-            committedEvent,
-            sourceType,
-            isFromCurrentActor,
-          });
-
-          if (hasCommittedId) {
-            const nextWatermark = Math.max(lastCommittedId, committedId);
-            await persistCommittedCursor(projectId, nextWatermark);
-          }
-        }),
-    });
-
-    await collabSession.start();
-    collabLog("info", "session started", {
-      projectId: resolvedProjectId,
-      mode,
-      partitions: resolvedPartitions,
-      online: Boolean(endpointUrl),
-    });
-    if (endpointUrl) {
-      collabLog("info", "attaching websocket transport", {
-        endpointUrl,
-      });
-      const transport = createWebSocketTransport({
-        url: endpointUrl,
-        label: "routevn.collab.web.transport",
-      });
-      try {
-        await collabSession.setOnlineTransport(transport);
-        collabLog("info", "websocket transport attached", {
-          endpointUrl,
-        });
-      } catch (error) {
-        collabLog(
-          "warn",
-          "failed to attach websocket transport; continuing in offline mode",
-          {
-            endpointUrl,
-            error: error?.message || "unknown",
-          },
-        );
-        return collabSession;
-      }
-
-      try {
-        const syncStartedAt = Date.now();
-        await collabSession.syncNow({
-          timeoutMs: INITIAL_REMOTE_SYNC_TIMEOUT_MS,
-        });
-        // Wait until repository projection has finished applying synced events.
-        await queueCollabApply(projectId, async () => {});
-        const repositoryEvents = Array.isArray(repository.getEvents?.())
-          ? repository.getEvents()
-          : [];
-        const repositoryEventSummary =
-          summarizeRepositoryEventsForSync(repositoryEvents);
-        collabLog("info", "repository event snapshot before replay", {
-          projectId: resolvedProjectId,
-          ...repositoryEventSummary,
-        });
-
-        const localCommittedCursor = await ensureCommittedIdLoaded(projectId);
-        const uncommittedEvents = toUncommittedSyncSubmissionEvents({
-          repositoryEvents,
-          committedCursor: localCommittedCursor,
-        });
-        collabLog("info", "uncommitted local event snapshot", {
-          projectId: resolvedProjectId,
-          localCommittedCursor,
-          localRepositoryEventCount: repositoryEvents.length,
-          uncommittedEventCount: uncommittedEvents.length,
-          firstUncommittedSchema:
-            uncommittedEvents[0]?.event?.payload?.schema || null,
-          firstUncommittedProjectId:
-            uncommittedEvents[0]?.event?.payload?.projectId || null,
-        });
-
-        if (uncommittedEvents.length > 0) {
-          for (const entry of uncommittedEvents) {
-            await collabSession.submitEvent({
-              partitions: entry.partitions,
-              event: entry.event,
-            });
-          }
-          await collabSession.flushDrafts();
-          await collabSession.syncNow({
-            timeoutMs: INITIAL_REMOTE_SYNC_TIMEOUT_MS,
-          });
-          await queueCollabApply(projectId, async () => {});
-          collabLog("info", "uncommitted local events submitted", {
-            projectId: resolvedProjectId,
-            submittedCount: uncommittedEvents.length,
-          });
-        }
-
-        const syncDurationMs = Date.now() - syncStartedAt;
-        collabLog("info", "initial remote sync completed", {
-          projectId: resolvedProjectId,
-          endpointUrl,
-          syncDurationMs,
-          repositoryEventCount: repositoryEvents.length,
-          replayedUncommittedEventCount: uncommittedEvents.length,
-        });
-      } catch (error) {
-        collabLog("warn", "initial remote sync failed", {
-          projectId: resolvedProjectId,
-          endpointUrl,
-          error: error?.message || "unknown",
-        });
-      }
-    }
-
-    return collabSession;
-  };
-
-  // Get or create repository by projectId
-  const getRepositoryByProject = async (projectId) => {
-    return getOrCreateLocked({
-      cache: repositoriesByProject,
-      locks: initLocksByProject,
-      key: projectId,
-      create: async () => {
-        const store = await getAdapterByProject(projectId);
-        const existingEvents = (await store.getEvents()) || [];
-
-        // Recover from stale sync cursor state: a project can have a persisted
-        // cursor while local repository events are empty/minimal (e.g. cleared IDB events).
-        const persistedCursor = await loadCommittedCursor({
-          adapter: store,
-          projectId,
-        }).catch(() => 0);
-        if (
-          persistedCursor > existingEvents.length &&
-          existingEvents.length <= 1
-        ) {
-          await clearCommittedCursor({
-            adapter: store,
-            projectId,
-          }).catch(() => {});
-          collabLastCommittedIdByProject.delete(projectId);
-          collabLog("warn", "reset stale committed cursor for project", {
-            projectId,
-            persistedCursor,
-            localEventCount: existingEvents.length,
-          });
-        }
-
-        const repository = await createProjectRepository({
-          projectId,
-          store,
-          events: existingEvents,
-        });
-
-        assertV2State(repository.getState());
-        repositoriesByProject.set(projectId, repository);
-        return repository;
-      },
-    });
-  };
-
-  // Get current project's repository (updates cache)
-  const getCurrentRepository = async () => {
-    const projectId = getCurrentProjectId();
-    if (!projectId) {
-      throw new Error("No project selected (missing ?p= in URL)");
-    }
-    const repository = await getRepositoryByProject(projectId);
-    // Update cache
-    currentRepository = repository;
-    currentProjectId = projectId;
-    currentAdapter = adaptersByProject.get(projectId);
-    return repository;
-  };
-
-  // Get cached repository (sync) - throws if not initialized
-  const getCachedRepository = () => {
-    const projectId = getCurrentProjectId();
-    if (!currentRepository || currentProjectId !== projectId) {
-      throw new Error(
-        "Repository not initialized. Call ensureRepository() first.",
-      );
-    }
-    return currentRepository;
-  };
-
-  const getCachedAdapter = () => {
-    const projectId = getCurrentProjectId();
-    if (!currentAdapter || currentProjectId !== projectId) {
-      throw new Error(
-        "Adapter not initialized. Call ensureRepository() first.",
-      );
-    }
-    return currentAdapter;
-  };
-
-  // File operations helpers
-  const storeFile = async (file) => {
-    const adapter = getCachedAdapter();
-    const fileId = nanoid();
-    const fileBlob = new Blob([await file.arrayBuffer()], { type: file.type });
-    await adapter.setFile(fileId, fileBlob);
-    return { fileId };
-  };
-
-  const getFileUrl = async (fileId) => {
-    const adapter = getCachedAdapter();
-    const blob = await adapter.getFile(fileId);
-    if (!blob) {
-      throw new Error(`File not found: ${fileId}`);
-    }
-    const url = URL.createObjectURL(blob);
-    return { url, type: blob.type };
-  };
-
-  const getBundleStaticFiles = async () => {
-    let indexHtml = null;
-    let mainJs = null;
-
-    try {
-      const indexResponse = await fetch("/bundle/index.html");
-      if (indexResponse.ok) {
-        indexHtml = await indexResponse.text();
-      }
-
-      const mainJsResponse = await fetch("/bundle/main.js");
-      if (mainJsResponse.ok) {
-        mainJs = await mainJsResponse.text();
-      }
-    } catch (error) {
-      console.error("Failed to fetch static bundle files:", error);
-    }
-
-    return { indexHtml, mainJs };
-  };
-
-  const storeMetadata = async (data) => {
-    const jsonString = JSON.stringify(data, null, 2);
-    const jsonBlob = new Blob([jsonString], { type: "application/json" });
-    const uniqueName = `metadata_${nanoid()}.json`;
-    Object.defineProperty(jsonBlob, "name", {
-      value: uniqueName,
-      writable: false,
-    });
-    return await storeFile(jsonBlob);
-  };
-
-  // File processors
-  const processors = {
-    image: async (file) => {
-      const dimensions = await getImageDimensions(file);
-      const { fileId } = await storeFile(file);
-      return { fileId, dimensions, type: "image" };
+    createStore: async ({ reference }) => {
+      return createInsiemeWebStoreAdapter(reference.projectId);
     },
-    audio: async (file) => {
-      const arrayBuffer = await file.arrayBuffer();
-      const fileForWaveform = new File([arrayBuffer], file.name, {
-        type: file.type,
-      });
-      const fileForStorage = new File([arrayBuffer], file.name, {
-        type: file.type,
-      });
 
-      const waveformData = await extractWaveformData(fileForWaveform);
-      const { fileId } = await storeFile(fileForStorage);
-
-      let waveformDataFileId = null;
-      if (waveformData) {
-        const compressedWaveformData = {
-          ...waveformData,
-          amplitudes: waveformData.amplitudes.map((value) =>
-            Math.round(value * 255),
-          ),
-        };
-        const waveformResult = await storeMetadata(compressedWaveformData);
-        waveformDataFileId = waveformResult.fileId;
-      }
-      return {
-        fileId,
-        waveformDataFileId,
-        waveformData,
-        duration: waveformData?.duration,
-        type: "audio",
-      };
-    },
-    video: async (file) => {
-      const [dimensions, thumbnailData] = await Promise.all([
-        getVideoDimensions(file),
-        extractVideoThumbnail(file, {
-          timeOffset: 1,
-          width: 240,
-          height: 135,
-          format: "image/jpeg",
-          quality: 0.8,
-        }),
-      ]);
-      const [videoResult, thumbnailResult] = await Promise.all([
-        storeFile(file),
-        storeFile(thumbnailData.blob),
-      ]);
-      return {
-        fileId: videoResult.fileId,
-        thumbnailFileId: thumbnailResult.fileId,
-        thumbnailData,
-        dimensions,
-        type: "video",
-      };
-    },
-    font: async (file) => {
-      const fontName = file.name.replace(/\.(ttf|otf|woff|woff2|ttc)$/i, "");
-      const fontUrl = URL.createObjectURL(file);
-
-      try {
-        await loadFont(fontName, fontUrl);
-      } catch (loadError) {
-        URL.revokeObjectURL(fontUrl);
-        throw new Error(`Invalid font file: ${loadError.message}`);
-      }
-
-      const { fileId } = await storeFile(file);
-      return { fileId, fontName, fontUrl, type: "font" };
-    },
-    generic: async (file) => {
-      const { fileId } = await storeFile(file);
-      return { fileId, type: "generic" };
-    },
-  };
-
-  const processFile = async (file) => {
-    const fileType = detectFileType(file);
-    const processor = processors[fileType] || processors.generic;
-    return processor(file);
-  };
-  const serviceCore = createProjectServiceCore({
-    router,
-    idGenerator: nanoid,
-    collabLog,
-    getCurrentRepository,
-    getCachedRepository,
-    getRepositoryByProject,
-    getAdapterByProject: (projectId) => adaptersByProject.get(projectId),
-    createSessionForProject,
-    createTransport: ({ endpointUrl }) =>
-      createWebSocketTransport({
-        url: endpointUrl,
-        label: "routevn.collab.web.transport",
-      }),
-    onEnsureLocalSession: ({ projectId }) => {
-      collabLog(
-        "warn",
-        "creating local-only session (backend transport not attached)",
-        {
-          projectId,
-          hint: "Set collabEndpoint or use default wss://127.0.0.1:8787/sync auto-connect.",
-        },
-      );
-    },
-    onSessionCleared: ({ projectId }) => {
-      collabApplyQueueByProject.delete(projectId);
-      collabAppliedCommittedIdsByProject.delete(projectId);
-    },
-    onSessionTransportUpdated: () => {},
-  });
-
-  return {
-    // Repository access - uses current project from URL
-    async getRepository() {
-      return getCurrentRepository();
-    },
-    async getRepositoryById(projectId) {
-      return getRepositoryByProject(projectId);
-    },
-    getAdapterById(projectId) {
-      return adaptersByProject.get(projectId);
-    },
-    async getFileByProjectId(projectId, fileId) {
-      const adapter = await getAdapterByProject(projectId);
-      return adapter.getFile(fileId);
-    },
-    async ensureRepository() {
-      return getCurrentRepository();
-    },
-    ...serviceCore.commandApi,
-    addVersionToProject: serviceCore.addVersionToProject,
-    deleteVersionFromProject: serviceCore.deleteVersionFromProject,
-    deleteResourceItemIfUnused: serviceCore.deleteResourceItemIfUnused,
-    async initializeProject({ projectId, template }) {
+    initializeProject: async ({ projectId, template }) => {
       return initializeWebProject({
         projectId,
         template,
       });
     },
-    createCollabSession: serviceCore.createCollabSession,
-    getCollabSession: serviceCore.getCollabSession,
-    getCollabSessionMode: serviceCore.getCollabSessionMode,
-    stopCollabSession: serviceCore.stopCollabSession,
-    submitCommand: serviceCore.submitCommand,
-    async uploadFiles(files) {
-      const fileArray = Array.isArray(files) ? files : Array.from(files);
-      const uploadPromises = fileArray.map(async (file) => {
-        try {
-          const result = await processFile(file);
-          return {
-            success: true,
-            file,
-            displayName: file.name.replace(/\.[^.]+$/, ""),
-            ...result,
-          };
-        } catch (error) {
-          console.error(`Failed to upload ${file.name}:`, error);
-          return { success: false, file, error: error.message };
-        }
+  };
+
+  const fileAdapter = {
+    continueOnUploadError: true,
+
+    storeFile: async ({ file, idGenerator, getCurrentStore }) => {
+      const adapter = getCurrentStore();
+      const fileId = idGenerator();
+      const fileBlob = new Blob([await file.arrayBuffer()], {
+        type: file.type,
       });
-      const results = await Promise.all(uploadPromises);
-      return results.filter((r) => r.success);
+      await adapter.setFile(fileId, fileBlob);
+      return { fileId };
     },
-    async getFileContent(fileId) {
-      return await getFileUrl(fileId);
-    },
-    async downloadMetadata(fileId) {
-      try {
-        const { url } = await getFileUrl(fileId);
-        const response = await fetch(url);
-        if (response.ok) {
-          const data = await response.json();
-          URL.revokeObjectURL(url);
-          return data;
-        }
-        URL.revokeObjectURL(url);
-        console.error("Failed to download metadata:", response.statusText);
-        return null;
-      } catch (error) {
-        console.error("Failed to download metadata:", error);
-        return null;
+
+    getFileContent: async ({ fileId, getCurrentStore }) => {
+      const adapter = getCurrentStore();
+      const blob = await adapter.getFile(fileId);
+      if (!blob) {
+        throw new Error(`File not found: ${fileId}`);
       }
+
+      const url = URL.createObjectURL(blob);
+      return {
+        url,
+        type: blob.type,
+        revoke: () => URL.revokeObjectURL(url),
+      };
     },
-    async loadFontFile({ fontName, fileId }) {
-      if (!fontName || !fileId || fileId === "undefined") {
-        throw new Error(
-          "Invalid font parameters: fontName and fileId are required.",
-        );
-      }
-      try {
-        const { url } = await getFileUrl(fileId);
-        await loadFont(fontName, url);
-        // Do not revoke here, font needs it
-        return { success: true };
-      } catch (error) {
-        console.error("Failed to load font file:", error);
-        return { success: false, error: error.message };
-      }
+
+    getFileByProjectId: async ({ projectId, fileId, getStoreByProject }) => {
+      const adapter = await getStoreByProject(projectId);
+      return adapter.getFile(fileId);
     },
-    detectFileType,
-    createBundle(projectData, assets) {
-      return createBundle(projectData, assets);
-    },
-    exportProject(projectData, files) {
-      return createBundle(projectData, files);
-    },
-    async downloadBundle(bundle, filename) {
+
+    downloadBundle: async ({ bundle, filename, filePicker }) => {
       await filePicker.saveFilePicker(
         new Blob([bundle], { type: "application/octet-stream" }),
         filename,
       );
-      return filename; // In web, we don't get a path back
+      return filename;
     },
-    async createDistributionZip(bundle, zipName) {
+
+    createDistributionZip: async ({
+      bundle,
+      zipName,
+      filePicker,
+      staticFiles,
+    }) => {
       const zip = new JSZip();
       zip.file("package.bin", bundle);
-
-      const { indexHtml, mainJs } = await getBundleStaticFiles();
-      if (indexHtml) zip.file("index.html", indexHtml);
-      if (mainJs) zip.file("main.js", mainJs);
+      if (staticFiles.indexHtml) zip.file("index.html", staticFiles.indexHtml);
+      if (staticFiles.mainJs) zip.file("main.js", staticFiles.mainJs);
 
       const zipBlob = await zip.generateAsync({ type: "blob" });
       await filePicker.saveFilePicker(zipBlob, `${zipName}.zip`);
       return `${zipName}.zip`;
     },
-    async createDistributionZipStreamed(projectData, fileIds, zipName) {
+
+    createDistributionZipStreamed: async ({
+      projectData,
+      fileIds,
+      zipName,
+      filePicker,
+      staticFiles,
+      getFileContent,
+    }) => {
       const uniqueFileIds = [];
       const seenFileIds = new Set();
 
@@ -893,21 +269,378 @@ export const createProjectService = ({
       const files = {};
       for (const fileId of uniqueFileIds) {
         try {
-          const content = await getFileUrl(fileId);
+          const content = await getFileContent(fileId);
           const response = await fetch(content.url);
           const buffer = await response.arrayBuffer();
           files[fileId] = {
             buffer: new Uint8Array(buffer),
             mime: content.type,
           };
-          URL.revokeObjectURL(content.url);
+          content.revoke?.();
         } catch (error) {
           console.warn(`Failed to fetch file ${fileId}:`, error);
         }
       }
 
       const bundle = await createBundle(projectData, files);
-      return await this.createDistributionZip(bundle, zipName);
+      return fileAdapter.createDistributionZip({
+        bundle,
+        zipName,
+        filePicker,
+        staticFiles,
+      });
     },
   };
+
+  const collabAdapter = {
+    beforeCreateRepository: async ({ projectId, store, events }) => {
+      const persistedCursor = await loadCommittedCursor({
+        adapter: store,
+        projectId,
+      }).catch(() => 0);
+
+      if (persistedCursor > events.length && events.length <= 1) {
+        await clearCommittedCursor({
+          adapter: store,
+          projectId,
+        }).catch(() => {});
+        collabLastCommittedIdByProject.delete(projectId);
+        collabLog("warn", "reset stale committed cursor for project", {
+          projectId,
+          persistedCursor,
+          localEventCount: events.length,
+        });
+      }
+
+      return events;
+    },
+
+    createTransport: ({ endpointUrl }) =>
+      createWebSocketTransport({
+        url: endpointUrl,
+        label: "routevn.collab.web.transport",
+      }),
+
+    onEnsureLocalSession: ({ projectId }) => {
+      collabLog(
+        "warn",
+        "creating local-only session (backend transport not attached)",
+        {
+          projectId,
+          hint: "Set collabEndpoint or use default wss://127.0.0.1:8787/sync auto-connect.",
+        },
+      );
+    },
+
+    onSessionCleared: ({ projectId }) => {
+      collabApplyQueueByProject.delete(projectId);
+      collabAppliedCommittedIdsByProject.delete(projectId);
+    },
+
+    onSessionTransportUpdated: () => {},
+
+    createSessionForProject: async ({
+      projectId,
+      token,
+      userId,
+      clientId,
+      endpointUrl,
+      partitions,
+      mode,
+      partitioning,
+      getRepositoryByProject,
+      getStoreByProject,
+      getProjectMetadataFromEntries,
+      collabLog,
+    }) => {
+      collabLog("info", "create session requested", {
+        projectId,
+        endpointUrl: endpointUrl || null,
+        mode,
+        hasToken: Boolean(token),
+        userId,
+        clientId,
+      });
+
+      const repository = await getRepositoryByProject(projectId);
+      const state = repository.getState();
+      assertV2State(state);
+      const adapter = await getStoreByProject(projectId);
+
+      const resolvedProjectId = state.project?.id || projectId;
+      const resolvedPartitions = partitioning.getBasePartitions(
+        resolvedProjectId,
+        partitions,
+      );
+      const projectMetadata = await getProjectMetadataFromEntries(projectId);
+      const clientStore = await createPersistedInMemoryClientStore({
+        projectId,
+        adapter,
+        logger: (entry) => {
+          if (entry?.level === "warn") {
+            collabLog("warn", "client store warning", entry);
+          }
+        },
+      });
+      await ensureCommittedIdLoaded(projectId, getStoreByProject);
+      const collabSession = createProjectCollabService({
+        projectId: resolvedProjectId,
+        projectName: projectMetadata.name,
+        projectDescription: projectMetadata.description,
+        initialState: projectRepositoryStateToDomainState({
+          repositoryState: state,
+          projectId: resolvedProjectId,
+        }),
+        token,
+        actor: {
+          userId,
+          clientId,
+        },
+        partitions: resolvedPartitions,
+        clientStore,
+        logger: (entry) => {
+          if (entry?.event === "connected") {
+            const globalLastCommittedId = Number(
+              entry?.global_last_committed_id,
+            );
+            if (Number.isFinite(globalLastCommittedId)) {
+              collabLog("debug", "connected cursor", {
+                projectId: resolvedProjectId,
+                globalLastCommittedId,
+              });
+            }
+          }
+          collabLog("debug", "sync-client", entry);
+        },
+        onCommittedCommand: ({
+          command,
+          committedEvent,
+          sourceType,
+          isFromCurrentActor,
+        }) =>
+          queueCollabApply(projectId, async () => {
+            const applyCommittedEventToRepository = async ({
+              command,
+              committedEvent,
+              sourceType,
+              isFromCurrentActor,
+            }) => {
+              if (isFromCurrentActor) {
+                collabLog("debug", "command skipped (current actor source)", {
+                  projectId,
+                  sourceType,
+                  commandId: command?.id || null,
+                  commandType: command?.type || null,
+                  committedId: Number.isFinite(
+                    Number(committedEvent?.committed_id),
+                  )
+                    ? Number(committedEvent.committed_id)
+                    : null,
+                });
+                return;
+              }
+
+              const beforeState = repository.getState();
+              const beforeImagesCount = countImageEntries(beforeState?.images);
+              const applyResult = await applyCommandToRepository({
+                repository,
+                command,
+                projectId: resolvedProjectId,
+              });
+              if (applyResult.events.length === 0) {
+                collabLog("warn", "command ignored (no projection events)", {
+                  projectId,
+                  sourceType,
+                  commandType: command?.type || null,
+                  commandId: command?.id || null,
+                });
+                return;
+              }
+
+              const afterState = repository.getState();
+              const afterImagesCount = countImageEntries(afterState?.images);
+              collabLog("info", "remote command applied to repository", {
+                projectId,
+                commandType: command?.type || null,
+                applyMode: applyResult.mode,
+                eventCount: applyResult.events.length,
+                beforeImagesCount,
+                afterImagesCount,
+                sourceType,
+              });
+              if (typeof onRemoteEvent === "function") {
+                for (const event of applyResult.events) {
+                  onRemoteEvent({
+                    projectId,
+                    sourceType,
+                    command,
+                    committedEvent,
+                    event,
+                  });
+                }
+              }
+            };
+
+            const committedId = Number(committedEvent?.committed_id);
+            const hasCommittedId = Number.isFinite(committedId);
+            const lastCommittedId = await ensureCommittedIdLoaded(
+              projectId,
+              getStoreByProject,
+            );
+
+            if (hasCommittedId) {
+              const appliedCommittedIds = getAppliedCommittedSet(projectId);
+              if (appliedCommittedIds.has(committedId)) {
+                collabLog(
+                  "debug",
+                  "committed event skipped (already applied)",
+                  {
+                    projectId,
+                    committedId,
+                    sourceType,
+                  },
+                );
+                return;
+              }
+              appliedCommittedIds.add(committedId);
+              if (appliedCommittedIds.size > 5000) {
+                const oldest = appliedCommittedIds.values().next().value;
+                appliedCommittedIds.delete(oldest);
+              }
+            }
+
+            await applyCommittedEventToRepository({
+              command,
+              committedEvent,
+              sourceType,
+              isFromCurrentActor,
+            });
+
+            if (hasCommittedId) {
+              const nextWatermark = Math.max(lastCommittedId, committedId);
+              await persistCommittedCursor(
+                projectId,
+                nextWatermark,
+                getStoreByProject,
+              );
+            }
+          }),
+      });
+
+      await collabSession.start();
+      collabLog("info", "session started", {
+        projectId: resolvedProjectId,
+        mode,
+        partitions: resolvedPartitions,
+        online: Boolean(endpointUrl),
+      });
+      if (endpointUrl) {
+        collabLog("info", "attaching websocket transport", {
+          endpointUrl,
+        });
+        const transport = createWebSocketTransport({
+          url: endpointUrl,
+          label: "routevn.collab.web.transport",
+        });
+        try {
+          await collabSession.setOnlineTransport(transport);
+          collabLog("info", "websocket transport attached", {
+            endpointUrl,
+          });
+        } catch (error) {
+          collabLog(
+            "warn",
+            "failed to attach websocket transport; continuing in offline mode",
+            {
+              endpointUrl,
+              error: error?.message || "unknown",
+            },
+          );
+          return collabSession;
+        }
+
+        try {
+          const syncStartedAt = Date.now();
+          await collabSession.syncNow({
+            timeoutMs: INITIAL_REMOTE_SYNC_TIMEOUT_MS,
+          });
+          await queueCollabApply(projectId, async () => {});
+          const repositoryEvents = Array.isArray(repository.getEvents?.())
+            ? repository.getEvents()
+            : [];
+          const repositoryEventSummary =
+            summarizeRepositoryEventsForSync(repositoryEvents);
+          collabLog("info", "repository event snapshot before replay", {
+            projectId: resolvedProjectId,
+            ...repositoryEventSummary,
+          });
+
+          const localCommittedCursor = await ensureCommittedIdLoaded(
+            projectId,
+            getStoreByProject,
+          );
+          const uncommittedEvents = toUncommittedSyncSubmissionEvents({
+            repositoryEvents,
+            committedCursor: localCommittedCursor,
+          });
+          collabLog("info", "uncommitted local event snapshot", {
+            projectId: resolvedProjectId,
+            localCommittedCursor,
+            localRepositoryEventCount: repositoryEvents.length,
+            uncommittedEventCount: uncommittedEvents.length,
+            firstUncommittedSchema:
+              uncommittedEvents[0]?.event?.payload?.schema || null,
+            firstUncommittedProjectId:
+              uncommittedEvents[0]?.event?.payload?.projectId || null,
+          });
+
+          if (uncommittedEvents.length > 0) {
+            for (const entry of uncommittedEvents) {
+              await collabSession.submitEvent({
+                partitions: entry.partitions,
+                event: entry.event,
+              });
+            }
+            await collabSession.flushDrafts();
+            await collabSession.syncNow({
+              timeoutMs: INITIAL_REMOTE_SYNC_TIMEOUT_MS,
+            });
+            await queueCollabApply(projectId, async () => {});
+            collabLog("info", "uncommitted local events submitted", {
+              projectId: resolvedProjectId,
+              submittedCount: uncommittedEvents.length,
+            });
+          }
+
+          const syncDurationMs = Date.now() - syncStartedAt;
+          collabLog("info", "initial remote sync completed", {
+            projectId: resolvedProjectId,
+            endpointUrl,
+            syncDurationMs,
+            repositoryEventCount: repositoryEvents.length,
+            replayedUncommittedEventCount: uncommittedEvents.length,
+          });
+        } catch (error) {
+          collabLog("warn", "initial remote sync failed", {
+            projectId: resolvedProjectId,
+            endpointUrl,
+            error: error?.message || "unknown",
+          });
+        }
+      }
+
+      return collabSession;
+    },
+  };
+
+  return createProjectServiceCore({
+    router,
+    db,
+    filePicker,
+    idGenerator: nanoid,
+    collabLog,
+    storageAdapter,
+    fileAdapter,
+    collabAdapter,
+  });
 };
